cambios finales2

+++ src/main/java/com/codarini/jpaservice/repository/DetalleFacturaRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface DetalleFacturaRepository extends JpaRepository<DetalleFacturaModel, Long> {
+public interface DetalleFacturaRepository extends JpaRepository<DetalleFacturaModel, Integer> {
    // List<Producto_VentaModel> findByIdProducto(int idProducto);
     //List<Producto_VentaModel> findByIdVenta(int idVenta);
     List<DetalleFacturaModel> findByCantidad(int cantidad);
